Split prescription page into details and track package tabs.
