+++ src/js/rx/routes.jsx
@@ -1,7 +1,9 @@
-import Active from './containers/Active';
-import Detail from './containers/Detail';
-import History from './containers/History';
 import Main from './containers/Main';
+import Active from './containers/Active';
+import History from './containers/History';
+import Prescription from './containers/Prescription';
+import Details from './containers/Details';
+import TrackPackage from './containers/TrackPackage';
 import Settings from './containers/Settings';
 import RxRefillsApp from './containers/RxRefillsApp';
 import GlossaryPage from './components/GlossaryPage';
@@ -20,7 +22,14 @@ const routes = {
     },
     { path: 'glossary', component: GlossaryPage },
     { path: 'settings', component: Settings },
-    { path: ':id', component: Detail }
+    {
+      path: ':id',
+      component: Prescription,
+      indexRoute: { component: Details },
+      childRoutes: [
+        { path: 'track', component: TrackPackage }
+      ]
+    }
   ]
 };
